requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
logging==0.4.9.6
multidict==6.1.0
propcache==0.2.0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
spotipy==2.24.0
urllib3==2.2.3
yarl==1.17.0
